# Speed up comparison page