@Kha 2023-08-05
> It's also a bit annoying how the table is not wide enough, even though there is enough space on my screen

Might get a bit *too* wide on 3440x1440 but I guess that's my own fault :)