@I-Al-Istannen 2023-08-05
> Awesome, thanks so much! What's wrong with `Map` though?

It can not handle object keys (or rather degrades to an identity map). I looked that up again, found that still to be the case, closed the tab and wrote the old-fashioned solution. I switched to a Map for the string keys now, that's probably still semantically nicer — even if still does not have proper keys.

> Might get a bit too wide on 3440x1440 but I guess that's my own fault :)

I vaguely recall having that exact conversation before :P You are using the software more often than I do, if it is an inconvenience for you that doesn't bode well for a change :) Sounds like the current behaviour is preferable then (or a different solution would be better, you are in a nicer position to judge that).

---

I hope that placates Keving Buzzard a bit, though I could also add a randomized "And the winner is" screen with a party emoji and confetti! Sounds like a good April Fools feature :P If he is still unhappy with the comparison, there might be something else wrong… Having that many dimensions wasn't quite something I had in mind, there very well may be some other place in there sorting invisible icons in quadratic time. But I would hate using software that keeps me waiting for a minute before showing a single god damn table, ironing those out should be fairly straightforward.